Generate schema for checkmate
